+++ src/middlewares/auth.middleware.js
@@ -1,7 +1,6 @@
 import { getUserFromId } from '#Repository/user.repository.js';
 import jwt from 'jsonwebtoken';
 import CustomError from '#Middleware/error/customError.js';
-import { getOwnerIdFromStoreId } from '#Repository/store.repository.js';
 
 //인증 미들웨어
 export const authMiddleware = async (req, res, next) => {
@@ -51,21 +50,4 @@ const user = await getUserFromId(decoded.userId);
   } catch (error) {
     return next();
   }
-};
-
-//가게 주인 확인
-export const checkStoreOwnership = async (req, res, next) => {
-    try {
-        const loggedInUserId = req.user.id;
-        const storeId = req.params.storeId;
-        
-        const ownerId = await getOwnerIdFromStoreId(storeId);
-
-        if (loggedInUserId !== ownerId) {
-            throw new CustomError({ name: 'FORBIDDEN' });
-        }
-        next();
-    } catch (err) {
-        next(err);
-    }
 };
