+++ src/middlewares/validation/user.validation.js
@@ -1,4 +1,5 @@
-import { body, query, param } from 'express-validator';
+import { body, param, query } from 'express-validator';
+import CustomError from '#Middleware/error/customError.js';
 import { email, password, refreshToken } from '#Util/pattern.js';
 
 //회원가입 검증
@@ -71,22 +72,6 @@ export const refreshTokenValidation = [
     .exists().withMessage('refreshToken 필드는 필수입니다.')
     .isString().withMessage('refreshToken은 문자열이어야 합니다.')
     .matches(refreshToken).withMessage('refreshToken 형식이 올바르지 않습니다.'),
-];
-
-// 내가 작성한 리뷰 목록 조회 커서 검증
-export const validateMyReviewsCursorQuery = [
-    query('cursor')
-        .optional()
-        .isString().withMessage('커서(cursor)는 문자열이어야 합니다.')
-        .matches(/^\d+$/).withMessage('커서(cursor)는 숫자 형식의 문자열이어야 합니다.')
-];
-
-//조회 제한 수 검증
-export const validateLimitQuery = [
-    query('limit')
-        .optional()
-        .isInt({ min: 1 }).withMessage('limit은 1 이상의 정수여야 합니다.')
-        .toInt()
 ];
 
 //유저 ID 파라미터 검증
