+++ src/repositories/user.repository.js
@@ -1,13 +1,10 @@
 import { prisma } from "../db.config.js";
 import { getRegionIdFromCode } from '#Repository/region.repository.js';
-import bcrypt from 'bcrypt';
 import CustomError from '#Middleware/error/customError.js';
-import { Prisma } from '@prisma/client';
 
 //유저 추가
 export const addUser = async (data) => {
   try {
-    const hashedPassword = await bcrypt.hash(data.password, 10);
 
     const newUser = await prisma.$transaction(async (tx) => {
       
@@ -27,7 +24,7 @@ export const addUser = async (data) => {
       const createdUser = await tx.user.create({
         data: {
           email: data.email,
-          password: hashedPassword,
+          password: data.hashedPassword,
           name: data.name,
           gender: data.gender,
           birthday: data.birthday,
