+++ src/services/user.service.js
@@ -1,18 +1,15 @@
-import {
-  addUser,
-  getUserIdPwFromEmail,
-  setfavoriteFood,
-  getProfileIdFromUserId
-} from "#Repository/user.repository.js";
-import { getReviewsByProfileId } from "#Repository/review.repository.js";
+import { addUser, getUserIdPwFromEmail, setfavoriteFood } from "#Repository/user.repository.js";
 import CustomError from "#Middleware/error/customError.js";
 import bcrypt from "bcrypt";
 
 //회원가입
 export const userSignUp = async (data) => {
+  
+  const hashedPassword = await bcrypt.hash(data.password, 10);
+
   const joinUserId = await addUser({
     email: data.email,
-    password: data.password,
+    hashedPassword,
     name: data.name,
     gender: data.gender,
     birthday: data.birthday,
@@ -43,28 +40,4 @@ export const userLogin = async (data) => {
   }
 
   return user.id;
-};
-
-//내가 작성한 리뷰 목록 조회
-export const getMyReviews = async (userId, cursor, requestedLimit) => {
-
-    const profileId = await getProfileIdFromUserId(userId);
-    if (!profileId) {
-        return { reviews: [], nextCursor: null, limit: requestedLimit || 10 }; 
-    }
-
-    let finalLimit = requestedLimit || 10;
-    if (finalLimit > 30) {
-        finalLimit = 30;
-    }
-
-    const reviews = await getReviewsByProfileId(profileId, cursor, finalLimit);
-
-    let nextCursor = null;
-    if (reviews.length === finalLimit) {
-        const lastReview = reviews[reviews.length - 1];
-        nextCursor = Number(lastReview.id);
-    }
-    
-    return { reviews, nextCursor, limit: finalLimit };
-};
+}
